# docs/devicerepository_swagger.yaml
basePath: /
definitions:
  api.AspectNodeQuery:
    properties:
      ids:
        items:
          type: string
        type: array
    type: object
...
  models.AspectNode:
    properties:
      ancestor_ids:
        items:
          type: string
        type: array
      child_ids:
        items:
          type: string
        type: array
      descendent_ids:
        items:
          type: string
        type: array
      id:
        type: string
      name:
        type: string
      parent_id:
        type: string
      root_id:
        type: string
    type: object
...
  models.Function:
    properties:
      concept_id:
        type: string
      description:
        type: string
      display_name:
        type: string
      id:
        type: string
      name:
        type: string
      rdf_type:
        type: string
    type: object
...
info:
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Device-Repository API
  version: "0.1"
paths:
  /aspect-nodes:
    get:
      deprecated: true
      description: deprecated list aspect-nodes; use GET /v2/aspect-nodes
      parameters:
      - description: filter; only 'measuring-function' is a valid value; if set, returns
          aspect-nodes used in combination with measuring-functions
        in: query
        name: function
        type: string
      - description: filter; in combination with 'function'; if true, returns also
          ancestor nodes of matching nodes
        in: query
        name: ancestors
        type: boolean
      - description: filter; in combination with 'function'; if true, returns also
          descendant nodes of matching nodes
        in: query
        name: descendants
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AspectNode'
            type: array
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: deprecated list aspect-nodes; use GET /v2/aspect-nodes
      tags:
      - aspect-nodes
  /aspect-nodes/{id}:
    get:
      description: get aspect-node
      parameters:
      - description: Aspect-Node Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AspectNode'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: get aspect-node
      tags:
      - aspect-nodes
  /aspect-nodes/{id}/measuring-functions:
    get:
      description: list measuring-functions used in combination with this aspect-node
      parameters:
      - description: Aspect-Node Id
        in: path
        name: id
        required: true
        type: string
      - description: filter; if true, returns also functions used in combination with
          ancestors of the input aspect-node
        in: query
        name: ancestors
        type: boolean
      - description: filter; if true, returns also functions used in combination with
          descendants of the input aspect-node
        in: query
        name: descendants
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Function'
            type: array
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: list aspect-node measuring-functions
      tags:
      - aspect-nodes
...
  /query/aspect-nodes:
    post:
      consumes:
      - application/json
      description: query aspect-nodes
      parameters:
      - description: AspectNodeQuery
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/api.AspectNodeQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AspectNode'
            type: array
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: query aspect-nodes
      tags:
      - aspect-nodes
...
  /v2/aspect-nodes:
    get:
      description: list aspect-nodes
      parameters:
      - description: default 100, will be ignored if 'ids' is set
        in: query
        name: limit
        type: integer
      - description: default 0, will be ignored if 'ids' is set
        in: query
        name: offset
        type: integer
      - description: filter
        in: query
        name: search
        type: string
      - description: default name.asc
        in: query
        name: sort
        type: string
      - description: filter; ignores limit/offset; comma-separated list
        in: query
        name: ids
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total-Count:
              description: count of all matching elements; used for pagination
              type: integer
          schema:
            items:
              $ref: '#/definitions/models.AspectNode'
            type: array
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - Bearer: []
      summary: list aspect-nodes
      tags:
      - aspect-nodes
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"